moved-components-to-RecipeDetailsPage-folder
