+++ myCookingManager/frontend/src/components/RecipeDetailsPage/Notepad.jsx
@@ -3,9 +3,9 @@ import { styled } from "styled-components";
 import { useAuth0 } from "@auth0/auth0-react";
 import { FaTimes } from "react-icons/fa";
 
-import { RecipeDetailsContext } from "./contexts/RecipeDetailsContext";
+import { RecipeDetailsContext } from "../contexts/RecipeDetailsContext";
 
-const DateTracker = () => {
+const Notepad = () => {
 	//Import user object from auth0
 	const { user } = useAuth0();
 
@@ -13,13 +13,13 @@ const DateTracker = () => {
 	const { currentRecipeDetails, setCurrentRecipeDetails } =
 		useContext(RecipeDetailsContext);
 
-	// States that store input onChange, dates in the array of dates and a flag to trigger the fetch
-	const [singleDate, setSingleDate] = useState("");
-	const [dates, setDates] = useState(currentRecipeDetails.dates_created);
-	const [wereDatesEdited, setWereDatesEdited] = useState(false);
+	// States that store input onChange, notes in the array of notes and a flag to trigger the fetch
+	const [singleNote, setSingleNote] = useState("");
+	const [notes, setNotes] = useState(currentRecipeDetails.notes);
+	const [wereNotesEdited, setWereNotesEdited] = useState(false);
 
 	useEffect(() => {
-		if (wereDatesEdited) {
+		if (wereNotesEdited) {
 			fetch(
 				`/api/user/${user.sub}/recipes/${currentRecipeDetails.recipeId}/update`,
 				{
@@ -29,7 +29,7 @@ const DateTracker = () => {
 						"Content-Type": "application/json",
 					},
 					body: JSON.stringify({
-						info: { dates_created: dates },
+						info: { notes: notes },
 					}),
 				}
 			)
@@ -40,9 +40,9 @@ const DateTracker = () => {
 						console.log(parsedResponse);
 						setCurrentRecipeDetails({
 							...currentRecipeDetails,
-							dates_created: dates,
+							notes: notes,
 						});
-						setWereDatesEdited(false);
+						setWereNotesEdited(false);
 					} else {
 						throw new Error(parsedResponse.message);
 					}
@@ -51,85 +51,68 @@ const DateTracker = () => {
 					console.error("Fetch error:", error);
 				});
 		}
-	}, [wereDatesEdited]);
+	}, [wereNotesEdited]);
 
-	const handleClearInputAndAddDate = () => {
-		if (singleDate.length > 0) {
-			setDates([...dates, { id: singleDate, text: singleDate }]);
-			setSingleDate("");
-			setWereDatesEdited(true);
+	const handleClearInputAndAddNote = () => {
+		if (singleNote.length > 0) {
+			setNotes([
+				...notes,
+				{ id: singleNote + notes.length, text: singleNote },
+			]);
+			setSingleNote("");
+			setWereNotesEdited(true);
 		}
 	};
 
 	// ------------------------------------------------------------ //
-	// Remove a date from the dates array by keeping all the ones
+	// Remove a note from the notes array by keeping all the ones
 	// that don't match the id of the one we want to delete.
 	// ------------------------------------------------------------ //
-	const handleDeleteDate = (deletedDate) => {
-		setDates((currentDates) => {
-			return currentDates.filter((currentDate) => {
-				return currentDate.id !== deletedDate.id;
+	const handleDeleteNote = (deletedNote) => {
+		setNotes((currentNotes) => {
+			return currentNotes.filter((currentNote) => {
+				return currentNote.id !== deletedNote.id;
 			});
 		});
-		setWereDatesEdited(true);
-	};
-
-	// Create today's date automatically and format it nicely.
-	const handleToday = () => {
-		const todayInMill = Date.now();
-		const today = new Date(todayInMill).toLocaleString("en-ca", {
-			weekday: "long",
-			month: "long",
-			year: "numeric",
-			day: "numeric",
-			hour: "numeric",
-			minute: "numeric",
-			hour12: false,
-		});
-
-		setDates([...dates, { id: today, text: today }]);
-		setWereDatesEdited(true);
+		setWereNotesEdited(true);
 	};
 
 	return (
 		<Wrapper>
-			<Label htmlFor="date">Days created</Label>
-			<ListParent>
-				{currentRecipeDetails.dates_created.map((date, index) => {
+			<Label htmlFor="note">Cooking notes</Label>
+			<ul>
+				{currentRecipeDetails.notes?.map((note, index) => {
 					return (
-						<ListChild key={date.id + index}>
-							<DateDiv>
-								<DateText>{date.text}</DateText>
-								{date.id && (
+						<li key={note.id + index}>
+							<NoteDiv>
+								<NoteText>{note.text}</NoteText>
+								{note.id && (
 									<DeleteButton
-										onClick={() => handleDeleteDate(date)}
+										onClick={() => handleDeleteNote(note)}
 									>
 										<DeleteIcon />
 									</DeleteButton>
 								)}
-							</DateDiv>
-						</ListChild>
+							</NoteDiv>
+						</li>
 					);
 				})}
-			</ListParent>
+			</ul>
 			<InputAndButtonDiv>
-				<ButtonDiv>
-					<TodayButton onClick={handleToday}>Today</TodayButton>
-					<Button
-						disabled={!singleDate.length}
-						onClick={() => handleClearInputAndAddDate()}
-					>
-						Add date manually
-					</Button>
-				</ButtonDiv>
-				<InputDiv>
-					<Input
-						id="date"
-						placeholder="Accepted formats: Sept 5 2023 or 02-05-2023 "
-						value={singleDate}
-						onChange={(ev) => setSingleDate(ev.target.value)}
-					/>
-				</InputDiv>
+				<Input
+					id="note"
+					type="text"
+					name="note"
+					placeholder="Write your note here !"
+					value={singleNote}
+					onChange={(ev) => setSingleNote(ev.target.value)}
+				/>
+				<Button
+					disabled={!singleNote.length}
+					onClick={() => handleClearInputAndAddNote()}
+				>
+					Add note
+				</Button>
 			</InputAndButtonDiv>
 		</Wrapper>
 	);
@@ -138,10 +121,9 @@ const DateTracker = () => {
 const Wrapper = styled.div`
 	margin: 20px 0;
 	padding: 5px 20px 10px;
-
+	border-radius: 5px;
 	box-shadow: 0 6px 20px 0 rgba(0, 0, 0, 0.19),
 		0 8px 30px 0 rgba(0, 0, 0, 0.18);
-	border-radius: 5px;
 `;
 
 const Label = styled.label`
@@ -152,27 +134,12 @@ const Label = styled.label`
 	font-weight: 700;
 `;
 
-const ListParent = styled.ul`
-	margin: 10px 0 10px 5px;
-`;
-
-const ListChild = styled.li`
-	width: 100%;
-	display: block;
-`;
-
-const InputDiv = styled.div`
-	display: flex;
-	justify-content: flex-start;
-	flex-grow: 1;
-`;
-
 const Input = styled.input`
 	padding: 6px;
 	width: 80%;
 	border: 2px solid var(--input-bg-color);
-	margin: 0 0 0 20px;
 	background-color: var(--input-bg-color);
+	font-family: var(--input-font-family);
 	font-size: 15px;
 
 	&:focus {
@@ -185,44 +152,21 @@ const InputAndButtonDiv = styled.div`
 	display: flex;
 	flex-direction: row;
 	justify-content: space-between;
-	align-items: flex-end;
-`;
-
-const ButtonDiv = styled.div`
-	display: flex;
-	flex-direction: column;
-	justify-content: center;
 	align-items: center;
 `;
 
 const Button = styled.button`
 	background-color: var(--tertiary-color);
-	color: black;
-	font-weight: 500;
 	border: 2px solid var(--tertiary-color);
-	padding: 5px;
-	width: 150px;
-	border-radius: 3px;
+	font-weight: 500;
 	font-size: 15px;
+	color: black;
+	padding: 5px;
+	min-width: 80px;
+	border-radius: 3px;
+	margin-top: 1px; // To be aligned with the input
 	opacity: ${(props) => (props.disabled ? "0.5" : "1")};
 	cursor: ${(props) => (props.disabled ? "default" : "pointer")};
-
-	&:focus {
-		border: 2px solid black;
-		outline: none;
-	}
-`;
-
-const TodayButton = styled.button`
-	background-color: var(--primary-color);
-	border: 2px solid var(--primary-color);
-	color: var(--secondary-color);
-	font-weight: 500;
-	padding: 5px;
-	width: 150px;
-	border-radius: 3px;
-	margin-bottom: 5px;
-	font-size: 15px;
 
 	&:focus {
 		border: 2px solid black;
@@ -248,17 +192,21 @@ const DeleteIcon = styled(FaTimes)`
 	margin-top: 1px;
 `;
 
-const DateDiv = styled.div`
+const NoteDiv = styled.div`
 	padding: 2px;
 	display: flex;
 	flex-direction: row;
 	justify-content: space-between;
 	align-items: center;
+	margin-left: 7px;
+	width: calc(
+		80% + (2 * 6px) + (2 * 2px)
+	); // To be the same length as the input field
 `;
 
-const DateText = styled.span`
+const NoteText = styled.span`
 	margin-right: 10px;
 	font-size: 15px;
 `;
 
-export default DateTracker;
+export default Notepad;
